requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
torchvision==0.14.1+cu117
openmim
mmengine>=0.7.1,<1.1.0
mmcv>=2.0.0rc4,<2.1.0
mmdet>=3.0.0rc5,<3.2.0
-e git+https://github.com/open-mmlab/mmocr.git#egg=mmocr
pyspellchecker
numpy
